Fix for #1737 - Better detect non default VirtualBox Installation Path
