+++ drivers/virtualbox/vbm.go
@@ -4,11 +4,8 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
-	"runtime"
 	"strings"
 
 	"github.com/docker/machine/libmachine/log"
@@ -77,32 +74,4 @@ func (v *VBoxCmdManager) vbmOutErr(args ...string) (string, string, error) {
 	}
 
 	return stdout.String(), stderrStr, err
-}
-
-// detectVBoxManageCmd detects the VBoxManage cmd's path if needed
-func detectVBoxManageCmd() string {
-	cmd := "VBoxManage"
-	if path, err := exec.LookPath(cmd); err == nil {
-		return path
-	}
-
-	if runtime.GOOS == "windows" {
-		if p := os.Getenv("VBOX_INSTALL_PATH"); p != "" {
-			if path, err := exec.LookPath(filepath.Join(p, cmd)); err == nil {
-				return path
-			}
-		}
-		if p := os.Getenv("VBOX_MSI_INSTALL_PATH"); p != "" {
-			if path, err := exec.LookPath(filepath.Join(p, cmd)); err == nil {
-				return path
-			}
-		}
-		// look at HKEY_LOCAL_MACHINE\SOFTWARE\Oracle\VirtualBox\InstallDir
-		p := "C:\\Program Files\\Oracle\\VirtualBox"
-		if path, err := exec.LookPath(filepath.Join(p, cmd)); err == nil {
-			return path
-		}
-	}
-
-	return cmd
 }
