+++ drivers/virtualbox/virtualbox_linux.go
@@ -5,3 +5,7 @@ package virtualbox
 func (d *Driver) IsVTXDisabled() bool {
 	return false
 }
+
+func detectVBoxManageCmd() string {
+	return detectVBoxManageCmdInPath()
+}
